+++ src/rendering/dx12/DX12Helpers.hpp
@@ -9,9 +9,9 @@ namespace Hyperion {
 enum ResourceState : uint8;
 enum TextureFormat : uint8;
 enum TextureType : uint8;
-enum FaceCullMode : uint32;
-enum Topology : uint32;
-enum BlendModeFactor : uint32;
+enum FaceCullMode : uint8;
+enum Topology : uint8;
+enum BlendModeFactor : uint8;
 enum StencilOp : uint8;
 enum StencilCompareOp : uint8;
